+++ src/index.ts
@@ -13,6 +13,8 @@ export * from './new-components/hidden-visually'
 export * from './new-components/alert'
 export * from './new-components/loading'
 export * from './new-components/notice'
+export * from './new-components/toast'
+export * from './new-components/notifications'
 
 // text and typography
 export * from './new-components/heading'