+++ src/new-components/notifications/index.ts
@@ -1,0 +1,2 @@
+export { useNotifications, NotificationsProvider } from './notifications'
+export type { NotificationProps, NotificationsProviderProps } from './notifications'