+++ src/new-components/toast/index.ts
@@ -1,0 +1,2 @@
+export { Toast } from './toast'
+export type { ToastProps } from './toast'